+++ docs/components/plugins/APISection.test.tsx
@@ -31,8 +31,8 @@ describe('APISection', () => {
     );
 
     expect(screen.getAllByRole('heading', { level: 2 })).toHaveLength(5);
-    expect(screen.getAllByRole('heading', { level: 3 })).toHaveLength(24);
-    expect(screen.getAllByRole('table')).toHaveLength(11);
+    expect(screen.getAllByRole('heading', { level: 3 })).toHaveLength(26);
+    expect(screen.getAllByRole('table')).toHaveLength(12);
 
     expect(screen.queryByText('Event Subscriptions'));
     expect(screen.queryByText('Components'));