+++ packages/expo-apple-authentication/ios/AppleAuthenticationModule.swift
@@ -3,21 +3,6 @@ import ExpoModulesCore
 
 let credentialRevokedEventName = "Expo.appleIdCredentialRevoked"
 
-struct FullName: Record {
-  @Field
-  var namePrefix: String?
-  @Field
-  var nameSuffix: String?
-  @Field
-  var givenName: String?
-  @Field
-  var middleName: String?
-  @Field
-  var familyName: String?
-  @Field
-  var nickname: String?
-}
-
 public final class AppleAuthenticationModule: Module {
   public func definition() -> ModuleDefinition {
     Name("ExpoAppleAuthentication")
@@ -48,7 +33,7 @@ public final class AppleAuthenticationModule: Module {
       }
     }
 
-    Function("formatFullName") { (fullName: FullName, formatStyle: FullNameFormatStyle?) in
+    Function("formatFullName") { (fullName: FullName, formatStyle: FullNameFormatStyle?) -> String in
       let formatStyle = formatStyle?.toFullNameFormatStyle() ?? .default
       var nameComponents = PersonNameComponents()