+++ packages/expo-camera/src/Camera.types.ts
@@ -354,6 +354,12 @@ export type CameraViewProps = ViewProps & {
    */
   videoQuality?: VideoQuality;
   /**
+   * The bitrate of the video recording in bits per second.
+   * Note: On iOS, you must specify the video codec when calling `recordAsync` to use this option.
+   * @example 10_000_000
+   */
+  videoBitrate?: number;
+  /**
    * A boolean that determines whether the camera shutter animation should be enabled.
    * @default true
    */