+++ packages/expo-splash-screen/src/index.native.ts
@@ -1,3 +1,4 @@
+import { isRunningInExpoGo } from 'expo';
 import { requireOptionalNativeModule } from 'expo-modules-core';
 
 import { SplashScreenNativeModule, SplashScreenOptions } from './SplashScreen.types';
@@ -6,6 +7,13 @@ const SplashModule = requireOptionalNativeModule<SplashScreenNativeModule>('Expo
 
 export function setOptions(options: SplashScreenOptions) {
   if (!SplashModule) {
+    return;
+  }
+
+  if (isRunningInExpoGo()) {
+    console.warn(
+      "'Splashscreen.setOptions' cannot be used in Expo Go. To customize the splash screen, you can use development builds."
+    );
     return;
   }