+++ packages/expo/src/winter/runtime.native.ts
@@ -1,5 +1,7 @@
 // This file configures the runtime environment to increase compatibility with WinterCG.
 // https://wintercg.org/
+
+// @ts-ignore: PolyfillFunctions does not have types exported
 import { polyfillGlobal as installGlobal } from 'react-native/Libraries/Utilities/PolyfillFunctions';
 
 import { installFormDataPatch } from './FormData';